Upgrade utils module to a package
